requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.5
gunicorn
dj-database-url
python-dotenv
whitenoise
# WebSocket support
channels==4.0.0
daphne==4.1.0
channels-redis==4.2.0

# Push notifications
pywebpush==2.0.0

